production/cassandra.yml: activeDeadlineSeconds for job and pod
Set activeDeadlineSeconds for the schema creation pod
to a little more than 5 minutes.

The schema creation code running in a container in this
pod already has a robust retry loop which should retry
forever. However, if that retry loop does not yield
success within ~5 minutes then restart the container
in the pod, effectively restarting the inner retry loop.
This guards against the unlikely case of the code running
in the container being dead-locked for whichever reason.

Set activeDeadlineSeconds on the job to 1 day. If the
retry loop in the container paired with the outer retrying
approach described in the previous paragraphs does not
result in success within 1 day (arbitrarily chosen)
then fail the job (this is a permanent failure state
requiring human intervention).

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>